Move API routes to separate class
